Move DEFAULT_DOMAIN constant to domains module
