+++ src/domains/helpers/DomainDropdown.tsx
@@ -9,9 +9,9 @@ import type { FC } from 'react';
 import { Link } from 'react-router-dom';
 import { DropdownItem } from 'reactstrap';
 import { useRoutesPrefix } from '../../utils/routesPrefix';
-import { DEFAULT_DOMAIN } from '../../visits/reducers/domainVisits';
 import { useVisitsComparisonContext } from '../../visits/visits-comparison/VisitsComparisonContext';
 import type { Domain } from '../data';
+import { DEFAULT_DOMAIN } from '../data';
 import type { EditDomainRedirects } from '../reducers/domainRedirects';
 import { EditDomainRedirectsModal } from './EditDomainRedirectsModal';
