+++ src/visits/reducers/domainVisits.ts
@@ -6,9 +6,6 @@ import type { LoadVisits, VisitsInfo } from './types';
 
 const REDUCER_PREFIX = 'shlink/domainVisits';
 
-// TODO Move to a generic module
-export const DEFAULT_DOMAIN = 'DEFAULT';
-
 interface WithDomain {
   domain: string;
 }
