+++ test/visits/visits-comparison/ShortUrlVisitsComparison.test.tsx
@@ -1,10 +1,10 @@
 import { cleanup, render, screen } from '@testing-library/react';
 import { fromPartial } from '@total-typescript/shoehorn';
 import { MemoryRouter } from 'react-router-dom';
+import { DEFAULT_DOMAIN } from '../../../src/domains/data';
 import type { MercureInfo } from '../../../src/mercure/reducers/mercureInfo';
 import type { ShortUrlIdentifier } from '../../../src/short-urls/data';
 import { queryToShortUrl, shortUrlToQuery } from '../../../src/short-urls/helpers';
-import { DEFAULT_DOMAIN } from '../../../src/visits/reducers/domainVisits';
 import { ShortUrlVisitsComparison } from '../../../src/visits/visits-comparison/ShortUrlVisitsComparison';
 import { checkAccessibility } from '../../__helpers__/accessibility';
